+++ robustnessgym/mosaic/columns/abstract.py
@@ -2,13 +2,15 @@ from __future__ import annotations
 
 import abc
 import logging
-from abc import abstractmethod
-from typing import Callable, Optional, Sequence
+from collections import defaultdict
+from typing import Callable, Dict, List, Mapping, Optional, Sequence, Union
 
 import numpy as np
 import torch
+from tqdm.auto import tqdm
 
 from robustnessgym.core.identifier import Identifier
+from robustnessgym.core.tools import convert_to_batch_column_fn
 from robustnessgym.mosaic.mixins.collate import CollateMixin
 from robustnessgym.mosaic.mixins.copying import CopyMixin
 from robustnessgym.mosaic.mixins.identifier import IdentifierMixin
@@ -17,6 +19,7 @@ from robustnessgym.mosaic.mixins.inspect_fn import FunctionInspectorMixin
 from robustnessgym.mosaic.mixins.materialize import MaterializationMixin
 from robustnessgym.mosaic.mixins.state import StateDictMixin
 from robustnessgym.mosaic.mixins.storage import ColumnStorageMixin
+from robustnessgym.mosaic.mixins.visibility import VisibilityMixin
 
 logger = logging.getLogger(__name__)
 
@@ -30,11 +33,11 @@ class AbstractColumn(
     IndexableMixin,
     MaterializationMixin,
     StateDictMixin,
+    VisibilityMixin,
     abc.ABC,
 ):
     """An abstract class for Mosaic columns."""
 
-    visible_rows: Optional[np.ndarray] = None
     _data: Sequence = None
 
     def __init__(
@@ -101,20 +104,6 @@ class AbstractColumn(
             # if not materializing, return a new Column
             return self.__class__(data, materialize=self.materialize)
 
-    def _remap_index(self, index):
-        if isinstance(index, int):
-            return self.visible_rows[index].item()
-        elif isinstance(index, slice):
-            return self.visible_rows[index].tolist()
-        elif isinstance(index, str):
-            return index
-        elif (isinstance(index, tuple) or isinstance(index, list)) and len(index):
-            return self.visible_rows[index].tolist()
-        elif isinstance(index, np.ndarray) and len(index.shape) == 1:
-            return self.visible_rows[index].tolist()
-        else:
-            raise TypeError("Invalid argument type: {}".format(type(index)))
-
     def __len__(self):
         # If only a subset of rows are visible
         if self.visible_rows is not None:
@@ -125,7 +114,6 @@ class AbstractColumn(
             return len(self.data)
         return 0
 
-    @abstractmethod
     def map(
         self,
         function: Optional[Callable] = None,
@@ -133,38 +121,129 @@ class AbstractColumn(
         batched: bool = False,
         batch_size: Optional[int] = 1000,
         drop_last_batch: bool = False,
+        num_proc: Optional[int] = None,
+        materialize: bool = None,
         **kwargs,
-    ) -> AbstractColumn:
+    ) -> Optional[Union[Dict, List, AbstractColumn]]:
         """Map a function over the elements of the column."""
-        raise NotImplementedError
+        # Check if need to materialize:
+        # TODO(karan): figure out if we need materialize=False
 
-    @abstractmethod
+        # Just return if the function is None
+        if function is None:
+            logger.info("`function` None, returning None.")
+            return None
+
+        # Ensure that num_proc is not None
+        if num_proc is None:
+            num_proc = 0
+
+        # Return if `self` has no examples
+        if not len(self):
+            logger.info("Dataset empty, returning None.")
+            return None
+
+        if not batched:
+            # Convert to a batch function
+            function = convert_to_batch_column_fn(function, with_indices=with_indices)
+            batched = True
+            logger.info(f"Converting `function` {function} to a batched function.")
+
+        # # Get some information about the function
+        # TODO: discuss whether this is actually required vs. doing it on first pass in
+        # loop
+        function_properties = self._inspect_function(
+            function,
+            with_indices,
+            batched=batched,
+        )
+
+        # Run the map
+        logger.info("Running `map`, the dataset will be left unchanged.")
+        outputs = defaultdict(list) if function_properties.dict_output else []
+        for i, batch in tqdm(
+            enumerate(
+                self.batch(
+                    batch_size=batch_size,
+                    drop_last_batch=drop_last_batch,
+                    collate=batched,
+                )
+            ),
+            total=(len(self) // batch_size)
+            + int(not drop_last_batch and len(self) % batch_size != 0),
+        ):
+
+            # Run `function` on the batch
+            output = (
+                function(
+                    batch,
+                    range(i * batch_size, min(len(self), (i + 1) * batch_size)),
+                )
+                if with_indices
+                else function(batch)
+            )
+
+            # Append the output
+            if output is not None:
+                if isinstance(output, Mapping):
+                    for k in output.keys():
+                        outputs[k].extend(output[k])
+                else:
+                    outputs.extend(output)
+
+        if not len(outputs):
+            return None
+        elif isinstance(outputs, dict):
+            # turns the defaultdict into dict
+            return dict(outputs)
+        return outputs
+
     def filter(
         self,
         function: Optional[Callable] = None,
-        with_indices: bool = False,
+        with_indices=False,
+        input_columns: Optional[Union[str, List[str]]] = None,
         batched: bool = False,
         batch_size: Optional[int] = 1000,
         drop_last_batch: bool = False,
+        num_proc: Optional[int] = 64,
         **kwargs,
-    ) -> AbstractColumn:
+    ) -> Optional[AbstractColumn]:
         """Filter the elements of the column using a function."""
-        raise NotImplementedError
+        # Just return if the function is None
+        if function is None:
+            logger.info("`function` None, returning None.")
+            return None
 
-    def set_visible_rows(self, indices: Optional[Sequence]):
-        """Set the visible rows of the column."""
-        if indices is None:
-            self.visible_rows = None
-        else:
-            if len(indices):
-                assert min(indices) >= 0 and max(indices) < len(self), (
-                    f"Ensure min index {min(indices)} >= 0 and "
-                    f"max index {max(indices)} < {len(self)}."
-                )
-            if self.visible_rows is not None:
-                self.visible_rows = self.visible_rows[np.array(indices, dtype=int)]
-            else:
-                self.visible_rows = np.array(indices, dtype=int)
+        # Return if `self` has no examples
+        if not len(self):
+            logger.info("Dataset empty, returning None.")
+            return None
+
+        # Get some information about the function
+        function_properties = self._inspect_function(
+            function,
+            with_indices,
+            batched=batched,
+        )
+        assert function_properties.bool_output, "function must return boolean."
+
+        # Map to get the boolean outputs and indices
+        logger.info("Running `filter`, a new dataset will be returned.")
+        outputs = self.map(
+            function=function,
+            with_indices=with_indices,
+            input_columns=input_columns,
+            batched=batched,
+            batch_size=batch_size,
+            drop_last_batch=drop_last_batch,
+            num_proc=num_proc,
+        )
+        indices = np.where(outputs)[0]
+
+        new_column = self.copy()
+        new_column.visible_rows = indices
+        return new_column
 
     def batch(
         self,
