+++ robustnessgym/mosaic/mixins/state.py
@@ -1,5 +1,5 @@
 from dataclasses import dataclass
-from typing import Collection, Dict
+from typing import Collection, Dict, Union
 
 from robustnessgym.core.tools import nested_map
 
@@ -25,12 +25,21 @@ class StateDictMixin:
     @classmethod
     def _assert_state_keys(cls, state: Dict) -> None:
         """Assert that a state contains all required keys."""
+        if cls._state_keys() is NotImplemented:
+            return
+
         assert (
             set(state.keys()) == cls._state_keys()
         ), f"State must contain all state keys: {cls._state_keys()}."
 
     def get_state(self) -> Dict:
-        """Get the internal state of the object."""
+        """
+        Get the internal state of the object.
+
+        For complex objects (e.g. Spacy Doc), this should
+        return a compressed representation of the object.
+
+        """
 
         def _apply_get_state(obj):
             if hasattr(obj, "get_state"):
@@ -38,16 +47,23 @@ class StateDictMixin:
             else:
                 return obj
 
-        state = nested_map(
-            _apply_get_state, {key: getattr(self, key) for key in self._state_keys()}
-        )
-        self._assert_state_keys(state)
+        if self._state_keys() is NotImplemented:
+            state = nested_map(
+                _apply_get_state,
+                {key: getattr(self, key) for key in self.__dict__.keys()},
+            )
+        else:
+            state = nested_map(
+                _apply_get_state,
+                {key: getattr(self, key) for key in self._state_keys()},
+            )
+            self._assert_state_keys(state)
+
         return state
 
     @classmethod
-    def from_state(cls, state: Dict, *args, **kwargs):
-        """Set the internal state of the dataset."""
-        cls._assert_state_keys(state)
+    def from_state(cls, state: Union[Dict, StateClass], *args, **kwargs) -> object:
+        """Load the object from state."""
 
         def _apply_from_state(obj_):
             if isinstance(obj_, StateClass):
@@ -55,7 +71,25 @@ class StateDictMixin:
             else:
                 return obj_
 
+        if isinstance(state, StateClass):
+            assert (
+                state.klass == cls
+            ), f"`state` has klass={state.klass} but `from_state` was called by {cls}."
+
+            # Extract the state dict
+            state = state.state
+
+        # Apply from state recursively
         state = nested_map(_apply_from_state, state)
-        obj = cls()
+
+        # Check that all keys are present
+        cls._assert_state_keys(state)
+
+        # Create a new object and update its state
+        try:
+            obj = cls(**state)
+        except TypeError:
+            obj = cls()
         obj.__dict__.update(state)
+
         return obj
